+++ scout/demo/resources/__init__.py
@@ -35,6 +35,10 @@ genes_to_phenotype_reduced_path = str(files(BASE_PATH).joinpath("genes_to_phenot
 phenotype_to_genes_reduced_path = str(files(BASE_PATH).joinpath("phenotype_to_genes_reduced.txt"))
 hpo_terms_def_path = str(files(BASE_PATH).joinpath("hpo_terms.csv"))
 
+# PanelApp path
+panelapp_all_reduced_path = str(files(BASE_PATH).joinpath("panelapp_panels_reduced.jsonl"))
+panelapp_all_ids_reduced_path = str(files(BASE_PATH).joinpath("panelapp_panel_ids_reduced.json"))
+
 demo_files = {
     "constraint_path": constraint_reduced_path,
     "genemap2_path": genemap2_reduced_path,