+++ scout/adapter/mongo/rank_model.py
@@ -60,6 +60,11 @@ class RankModelHandler(object):
 
         return {}
 
+    def delete_rank_model(self, rank_model_url: str) -> dict:
+        """Delete a rank model from database with URL as identifier"""
+
+        return self.rank_model_collection.delete_one({"_id": rank_model_url})
+
     def rank_model_url_from_version(
         self, rank_model_link_prefix: str, rank_model_version: str, rank_model_file_extension: str
     ) -> str: