+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+## [unreleased]
+### Added
+- Command to download all PanelApp panels to file ()
+
 ## [4.110.0]
 ### Added
 - A documentation page for institute settings (#6212)