+++ CHANGELOG.md
@@ -7,6 +7,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [unreleased]
 ### Added
 - Custom MT Saltshaker HTML report loading (#6242)
+- Command to download all PanelApp panels to file (#6239)
 
 ## [4.110.0]
 ### Added