+++ CHANGELOG.md
@@ -11,6 +11,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Custom MT Saltshaker HTML report loading (#6242)
 - Command to download all PanelApp panels to file (#6239)
 - Alternatively load and update PANELAPP-GREEN panel from a pre-downloaded file (#6244)
+### Changed
+- Default gnomAD linkout for hg38 to non_uk_biobank subset (#6247)
 ### Fixed
 - Saltshaker report update command (#6246)