+++ CHANGELOG.md
@@ -15,6 +15,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Default gnomAD linkout for hg38 to non_uk_biobank subset (#6247)
 ### Fixed
 - Saltshaker report update command (#6246)
+- Managed variants counter, wrongly defaulting to build 37 (#6252)
 
 ## [4.110.0]
 ### Added