+++ CHANGELOG.md
@@ -11,9 +11,11 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Alternatively load and update PANELAPP-GREEN panel from a pre-downloaded file (#6244)
 ### Changed
 - Default gnomAD linkout for hg38 to non_uk_biobank subset (#6247)
+- Handle conflicts and duplicates when uploading managed variants (#6256)
 ### Fixed
 - Saltshaker report update command (#6246)
 - Managed variants counter, wrongly defaulting to build 37 (#6252)
+- Managed variants matching follows build on case page (#6254)
 
 ## [4.110.0]
 ### Added